Fix order of how diffs are applied to basefile
The client used version items from "history" from project info in the order
as the server returns them... which is likely sorted, but using lexical sorting,
so in some cases it would fail - e.g. when updating from v4 to v11 the order
of diffs was v10, v11, v5, v6, v7, v8, v9. This makes sure that the versions
are always correctly ordered.

Fixes lutraconsulting/qgis-mergin-plugin#219